backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
python-multipart==0.0.9
python-docx==1.1.2
pypdf==4.2.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
